refactor(kv): extract `bind_params` and `pg_type_metadata` modules to separate files
